refactor: extract frame_compare CLI helpers into cli_utils
